feat: Add new operation to get a restaurant by its identifier (#12)
* Start using semver comments in dependabot

* Added new parameters to avoid exit on timeout

* Add offset and limit as flags in the proper cobra command

* Improvements in the dashboard also adding a new panel to monitor restaurant creation in database

* Simplify infra scripts

* Tidy and sync modules

* Automate grafana dashboard import when creating the infra

* Skip resource creation in Grafana for stop/start operations

* feat: Allow get a restaurant by its id

+++ internal/adapter/primary/api/rest/mapper.go
@@ -80,7 +80,7 @@ func (dm DefaultMapper) toDomainRestaurants(restaurants []*Restaurant) []*domain
 // FromDomainRestaurant maps a domain.Restaurant struct into a Restaurant.
 func (dm DefaultMapper) fromDomainRestaurant(r *domain.Restaurant) *Restaurant {
 	restRestaurant := Restaurant{}
-	restRestaurant.Id = int64(r.Id)
+	restRestaurant.Id = r.Id
 	restRestaurant.Name = r.Name
 	restRestaurant.Address = dm.fromDomainAddress(r.Address)
 	restRestaurant.Menu = dm.fromDomainMenu(r.Menu)
